Turn MossberryCost into CollectableItemCost
