+++ ItemChanger.Silksong/RawData/BaseLocationList/Tools.cs
@@ -13,11 +13,11 @@ internal static partial class BaseLocationList
     {
         Name = LocationNames.Druid_s_Eye,
         SceneName = SceneNames.Mosstown_02c,
-    }.WithTag(new ImplicitCostTag { Cost = new MossberryCost { Amount = 3 }, Inherent = true });
+    }.WithTag(new ImplicitCostTag { Cost = CollectableItemCost.Mossberries(3), Inherent = true });
 
     public static Location Druid_s_Eyes => new MossDruidTool2Location
     {
         Name = LocationNames.Druid_s_Eyes,
         SceneName = SceneNames.Mosstown_02c,
-    }.WithTag(new ImplicitCostTag { Cost = new MossberryCost { Amount = 1 }, Inherent = true });
+    }.WithTag(new ImplicitCostTag { Cost = CollectableItemCost.Mossberries(1), Inherent = true });
 }
